Store AGPS data to file on SIGUSR1
When the app received SIGUSR1, AGPS data is stored. This allows
periodic or on-demand storing of AGPS data.

A flag var is used to indicate when store() should be called in the main
loop. The reason for this is that awaiting in the signal handler
prevents any additional signals from being received, and also caused
some deadlock when a client was connected, likely from reads happening
for servicing the client and for the store operation. This flag in the
main loop seems to be a simple way to work around that.