gnss/stm: make device reading thread safe
This uses a mutex and a new 'readline' method to control read access to
the underlying stm device (both for serial and gnss), and resolves race
conditions that existed when multiple goroutines were all using separate
bufio.Scanners concurrently.

A more elegant way to implement this might be using 'context', or a
WaitGroup.